F - Flytter opprettelse av journalpost og oppgave til JournalforingService
